Pilko KoskiRequestMapper YKI- ja VKT-mappereihin
KoskiRequestMapper sisälsi kaksi noin sadan rivin sisäkkäistä builder-
metodia (yksi YKI:lle, yksi VKT:lle) ja per-kutsu rakennetun JSON-
mapperin. Splittaa kahteen @Service-luokkaan: KoskiYkiRequestMapper ja
KoskiVktRequestMapper, jolloin näkymäkontrollerit voivat injektoida vain
oman domeeninsa mapperin. KoskiService injektoi molemmat.

Lisäksi getObjectMapper()-companion-staticista tehdään Spring-bean
"koskiObjectMapper" omaan KoskiJacksonConfig-konfiguraatioluokkaansa.
Tämä välttää sykliriippuvuuden, joka syntyisi KoskiRestClientConfigin
sisällä (yhteystiedotJacksonConverter → koskiObjectMapper →
KoskiRestClientConfig → restClientBuilder → ...).

Spring rakentaa mapperin nyt kerran sovelluksen elinkaaren aikana, kun
aiemmin se rakennettiin uusiksi jokaisella JSON-renderöinnillä.
