+++ server/src/main/kotlin/fi/oph/kitu/koski/KoskiRestClientConfig.kt
@@ -1,10 +1,12 @@
 package fi.oph.kitu.koski
 
+import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.http.converter.json.JacksonJsonHttpMessageConverter
 import org.springframework.web.client.RestClient
+import tools.jackson.databind.json.JsonMapper
 import java.util.Base64
 
 @Configuration
@@ -21,14 +23,16 @@ class KoskiRestClientConfig(
     private lateinit var password: String
 
     @Bean("koskiRestClient")
-    fun restClient(): RestClient {
+    fun restClient(
+        @Qualifier("koskiObjectMapper") koskiObjectMapper: JsonMapper,
+    ): RestClient {
         val basicAuthToken = Base64.getEncoder().encodeToString("$user:$password".toByteArray())
         return restClientBuilder
             .baseUrl(koskiBaseUrl)
             .defaultHeader("Authorization", "Basic $basicAuthToken")
             .configureMessageConverters { cs ->
                 cs.registerDefaults().withJsonConverter(
-                    JacksonJsonHttpMessageConverter(KoskiRequestMapper.getObjectMapper()),
+                    JacksonJsonHttpMessageConverter(koskiObjectMapper),
                 )
             }.build()
     }
